requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
pandas==2.2.1
faker==24.0.0
numpy==1.26.4
pydantic==2.6.4
pydantic-settings==2.2.1
pytest==8.1.1
pytest-postgresql==6.0.0
streamlit==1.32.2
plotly==5.20.0
pyyaml==6.0.1
